Create README.md
How to run the project
To run this project, you should add images need to be processed to the input folder first. Please be caution that images should be RGB image and in order. Then run the stitchMain.m, you can see the output images after processing completely and the output image will be saved on output file folder.